Added 5 Extra Tests
As a part of the assignment, we had to make our own tests. The last 5 are the ones I made so if you have suggestions on how to make better tests please do so! Additionally, there are 3 tests that will not pass because they consider Optional Attendees. If there was extra time I could have added the element to my code (I don't have that time).